requirements.txt
PySide6>=6.6.0
pyswisseph>=2.10.0.0
pytz>=2024.1
geonamescache>=2.0.0
timezonefinder>=6.2.0
requests>=2.31.0
nuitka>=2.0
